+++ glider_dac/models/deployment.py
@@ -4,10 +4,12 @@
 glider_dac/models/deployment.py
 Model definition for a Deployment
 '''
-from glider_dac import app, db, slugify
+from glider_dac import app, db, slugify, queue
+from glider_dac.glider_emails import glider_deployment_check
 from datetime import datetime
 from flask_mongokit import Document
 from bson.objectid import ObjectId
+from rq import Queue, Connection, Worker
 from shutil import rmtree
 import os
 import hashlib
@@ -166,6 +168,7 @@ class Deployment(Document):
         - write or remove complete.txt
         - generate/update md5 files (removed on not-complete)
         - link/unlink via bindfs to archive dir
+        - schedule checker report against ERDDAP endpoint
         """
         # Save a file called "completed.txt"
         completed_file = os.path.join(self.full_path, "completed.txt")
@@ -196,11 +199,24 @@ class Deployment(Document):
                     md5_file = full_file + ".md5"
                     with open(md5_file, 'w') as mf:
                         mf.write(md5_value)
+            # schedule the checker job to kick off the compliance checker email
+            # on the deployment when the deployment is completed
+            # on_complete might be a misleading function name -- this section
+            # can run any time there is a sync, so check if a checker run has already been executed
+            if getattr(self, "compliance_check_passed", None) is None:
+                # eliminate force/always re-run?
+                app.logger.info("Scheduling compliance check for completed "
+                                "deployment {}".format(self.deployment_dir))
+                queue.enqueue(glider_deployment_check,
+                              kwargs={"deployment_dir": self.deployment_dir},
+                              job_timeout=800)
         else:
             for dirpath, dirnames, filenames in os.walk(self.full_path):
                 for f in filenames:
                     if f.endswith(".md5"):
                         os.unlink(os.path.join(dirpath, f))
+
+
 
     def calculate_checksum(self):
         '''
