+++ requirements.txt
@@ -26,5 +26,5 @@ rq-dashboard==0.3.4
 netCDF4>=1.4.0
 cf-units>=2
 flask-cors==3.0.2
-compliance-checker>=4.3.1
-cc-plugin-glider>=2.0.0
+compliance-checker>=4.3.3
+cc-plugin-glider>=2.0.3
